cleanup : improved tests organization in test directory some more
